Refactor code to reduce complexity
